Add code highlighting support for doxygen inline code.
Doxygen parses inline code and produces an annotated XML output with
semantical hihghlighting. It does not reveal the source language though.
This is a problem for breathe which can not forward the source code to
pygments.

Instead, this commit let's breathe produce html output directly while
re-using existing pygments styles. This will probably not work for
LaTeX output.